pyproject.toml
[project]
  name = "inyeon"
  version = "0.1.0"
  description = "Git workflow AI assistant"
  requires-python = ">=3.11"
  dependencies = [
      # CLI
      "typer>=0.12.0",
      "rich>=13.0.0",

      # HTTP
      "httpx>=0.27.0",

      # Validation
      "pydantic>=2.0.0",
      "pydantic-settings>=2.0.0",

      # Backend
      "fastapi>=0.110.0",
      "uvicorn[standard]>=0.29.0",

      # Diff parsing
      "unidiff>=0.7.5",

      # Agents
      "langgraph>=0.2.0",
      "langchain-core>=0.3.0",

      # LLM Providers
      "google-genai>=1.0.0",

      # RAG
      "chromadb>=0.4.0",
  ]

  [project.scripts]
  inyeon = "cli.main:app"

  [project.optional-dependencies]
  dev = [
      "pytest>=8.0.0",
      "pytest-asyncio>=0.23.0",
      "ruff>=0.3.0",
  ]

  [build-system]
  requires = ["setuptools>=61.0"]
  build-backend = "setuptools.build_meta"

  [tool.setuptools.packages.find]
  include = ["cli*", "backend*"]